requirements/common.txt
aiohttp==1.2.0
aiopg==0.13.0
async-timeout==1.1.0
chardet==2.3.0
cliff==2.4.0
cmd2==0.6.9
colorlog==2.10.0
ecdsa==0.13
future==0.16.0
jsonschema==2.5.1
multidict==2.1.4
pbr==1.10.0
prettytable==0.7.2
psycopg2==2.6.2
pycrypto==2.6.1
pyflakes==1.5.0
pyparsing==2.1.10
python-jose==1.3.2
PyYAML==3.12
requests==2.12.4
six==1.10.0
stevedore==1.19.1
structlog==16.1.0
yarl==0.8.1
